Split scope classes into files
