+++ lib/mongoid/report.rb
@@ -5,6 +5,9 @@ require_relative 'report/config'
 require_relative 'report/queries_builder'
 require_relative 'report/attach_proxy'
 require_relative 'report/collection'
+require_relative 'report/batches'
+require_relative 'report/merger'
+require_relative 'report/output'
 require_relative 'report/scope'
 require_relative 'report/scope_collection'
 require_relative 'report/report_proxy'
